==> code/2 check function.ipynb <==
def helix(filename):
    f = open(filename)
    count = 0
    for i in f:
        if "HELIX  " in i:
            count += 1
            
    return count


def sheet(filename):
    f = open(filename)
    count = 0
    for i in f:
        if "SHEET   " in i:
            count += 1
            
    return count

def check(filename):
    print(sheet(filename), "sheets")
    print(helix(filename), "helixes")

check("1l2y.pdb")

print(helix("2c0k.pdb"))
print(sheet("2c0k.pdb"))

print(helix("3q9h.pdb"))
print(sheet("3q9h.pdb"))

print(sheet("1q82.pdb"))
print(helix("1q82.pdb"))

print(sheet("1p3u.pdb1"))
print(helix("1p3u.pdb1"))

print(sheet("1luq.pdb"))
print(helix("1luq.pdb"))

def write_csv(PDB_list):
    g = open("PDB_alpha_beta","w")
    writer=csv.writer(g)
    
    header = ["ID","ALPHA","BETA"]
    writer.writerow(header)
    for i in PDB_list:
        data = [i,helix[i],sheet[i]]
        writer.writerow(data)
        
    g.close()

from Bio.PDB import *

p = PDBParser()
structure = p.get_structure("1l2y", "1l2y.pdb")

residues =[]
atoms_n = []
atoms_o = []

for model in structure:
    for chain in model:
        for residue in chain:
            for atom in residue:
                if atom.get_name()=='N':
                    atoms_n.append(atom)
                if atom.get_name()=="O":
                    atoms_o.append(atom)


#dir(structure)
# structure.get_residues()

atoms_o[0].get_vector()
