==> code/4 ML Model 1.ipynb <==
# WORKING ML MODEL TO PREDICT BETA SHEET USING NAIVE BAYES CLASSIFICATION

import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.naive_bayes import MultinomialNB
from sklearn.metrics import confusion_matrix, accuracy_score

def predict_beta(dataset):

    y = dataset.sheets
    X = dataset.sequences

    X_train, X_test, y_train, y_test=train_test_split(X,y,test_size=0.2)


    #to change from words to numbers

    vect = CountVectorizer()
    counts = vect.fit_transform(X_train.values)

    classifier = MultinomialNB() #this means it's a NB model
    targets=y_train.values  #need to put  (.value) because otherwise takes into consideration the indeces  
    classifier.fit(counts,targets)

    counts_test=vect.transform(X_test.values)
    y_pred =classifier.predict(counts_test)

    df=pd.DataFrame({'Actual':y_test,'Predicted':y_pred})

    #print(df)


    #mat=confusion_matrix(y_test,y_pred)
    #print(mat)
    #print('Accuracy', accuracy_score(y_test,y_pred))
    return accuracy_score(y_test, y_pred)

scores = []

dataset = pd.read_csv('training_data.csv')

for i in range(500): # running the model 500 times
    scores.append(predict_beta(dataset))
    
print("Average accuracy score", sum(scores)/500)
print("Maximum accuracy score:",max(scores))
print("Minimum accuracy score:", min(scores))


# predicting alpha helix


import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.naive_bayes import MultinomialNB
from sklearn.metrics import confusion_matrix, accuracy_score

def predict_alpha():
    dataset = pd.read_csv('Data/Old/proteins.csv')
    dataset.head()

    y = dataset.helices
    X = dataset.sequences

    length = dataset.length
    chains = dataset.chains

    X_train, X_test, y_train, y_test=train_test_split(X,y,test_size=0.2)


    #to change from words to numbers

    vect = CountVectorizer()
    counts = vect.fit_transform(X_train.values)

    classifier = MultinomialNB() #this means it's a NB model
    targets=y_train.values  #need to put  (.value) because otherwise takes into consideration the indeces  
    classifier.fit(counts,targets)

    counts_test=vect.transform(X_test.values)
    y_pred =classifier.predict(counts_test)

    df=pd.DataFrame({'Actual':y_test,'Predicted':y_pred})

    #print(df)


    #mat=confusion_matrix(y_test,y_pred)
    #print(mat)
    #print('Accuracy', accuracy_score(y_test,y_pred))
    return accuracy_score(y_test, y_pred)

scores = []

for i in range(500):
    scores.append(predict_alpha())
    
print("Average accuracy score", sum(scores)/500)
print(max(scores))
print(min(scores))

# Logistic regression using chains as the feature

import pandas as pd

dataset = pd.read_csv('Data/natural_seq.csv')
dataset = dataset[:2000]
dataset.shape

from sklearn.linear_model import LogisticRegression
from sklearn.model_selection import train_test_split
from sklearn.metrics import confusion_matrix, accuracy_score
from sklearn.model_selection import cross_validate
import numpy as np

X = pd.DataFrame()
X['x1'] = dataset["length"].values
X['x2'] = dataset["chains"].values

y = dataset.sheets

#X, y = dataset["length"].values.reshape(-1, 1), dataset.sheets
#X, y = [dataset["length"].values, dataset["chains"].values], dataset.sheets
                                   
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2)

clf = LogisticRegression(random_state=0).fit(X_train, y_train)

#y_pred = clf.predict(X_test)
#print(accuracy_score(y_pred, y_test))

cv_results = cross_validate(clf, X, y, cv=100)


print(np.average(cv_results["test_score"]))
