==> code/5 more data collection.ipynb <==
import pandas as pd

df = pd.read_csv('proteins_under_100.csv')
#df = pd.read_csv('proteins_many.csv')

df.head()

# d3to1 = {'CYS': 'C', 'ASP': 'D', 'SER': 'S', 'GLN': 'Q', 'LYS': 'K',
# 'ILE': 'I', 'PRO': 'P', 'THR': 'T', 'PHE': 'F', 'ASN': 'N', 
# 'GLY': 'G', 'HIS': 'H', 'LEU': 'L', 'ARG': 'R', 'TRP': 'W', 
# 'ALA': 'A', 'VAL':'V', 'GLU': 'E', 'TYR': 'Y', 'MET': 'M'}

# for i in d3to1:
#     proteins.append(i)

def get_natural_proteins(df):
    acids = ['CYS', 'ASP', 'SER', 'GLN', 'LYS', 'ILE', 'PRO', 'THR', 'PHE', 'ASN', 'GLY', 'HIS', 'LEU', 'ARG', 'TRP', 'ALA', 'VAL', 'GLU', 'TYR', 'MET']

    sequences = []
    index = []
    counter = 0

    for sequence in df.sequences:
        sequence = sequence.replace("'", "")
        sequence = sequence.replace(" ", "")
        sequence = sequence.replace("[", "")
        sequence = sequence.replace("]", "")
        sequence = sequence.split(",")
        for acid in sequence:
            if acid not in acids:
                break
        else:
            sequences.append(sequence)
            index.append(counter)
        counter += 1
    
    return sequences, index


natural = get_natural_proteins(df)

sequences = natural[0]
indices = natural[1]

#indices[-1]

#print(sequences[4])
#print(df.sequences[372]) # it works


natural_proteins = df.loc[indices]
natural_proteins.drop("Unnamed: 0", axis=1, inplace=True)
natural_proteins.to_csv("natural_proteins_under_100.csv")   # create csv


natural_proteins = pd.read_csv("natural_proteins_under_100.csv")  # read csv
natural_proteins.head()
