==> code/6 ML Model 2.ipynb <==
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.linear_model import LogisticRegression
from sklearn.model_selection import cross_validate
from sklearn.metrics import accuracy_score, confusion_matrix, r2_score
from sklearn import tree  # trying decision tree classifier
from sklearn import svm 

df = pd.read_csv("natural_proteins.csv")
df = df[df.chains ==1]  

# sheets = df.sheets[df.length <=1000]
# helices = df.helices[df.length <=1000]
# length = df.length[df.length <=1000]
# chains = df.chains[df.length <=1000]
# sequences = df.sequences[df.length <=1000]

names = df.names
sheets = df.sheets
helices = df.helices
chains = df.chains
sequences = df.sequences

converted = []
  
for i in sequences:
    # Converting string to list
    conv = i.strip('][').strip("'")#.split(', ')
    converted.append(conv)

sequences = converted

df = pd.DataFrame(list(zip(names, sequences, chains, helices, sheets)), columns = ["names", "sequences", "chains", "helices", "sheets"])
df.to_csv("1chain_proteins.csv")   # create csv

df = pd.read_csv("1chain_proteins.csv")  # read csv
df.head()


# Function that assigns numerical values to the acids in the sequences

import numpy as np
            
d3to1 = {'C': 0,'D': 1, 'S': 2, 'Q': 3, 'K': 4,
 'I': 5, 'P': 6, 'T': 7, 'F': 8, 'N': 9, 
 'G': 10, 'H': 11, 'L': 12, 'R': 13, 'W': 14, 
 'A': 15, 'V': 16, 'E': 17, 'Y': 18, 'M': 19, None: 20}

def sequence_enhancer(sequence, n):
    sequence_numeric_values = []
    
    for i in sequence:
        sequence_numeric_values.append(d3to1[i])
                
    if len(sequence) < n:
        for i in range(n-len(sequence)):
            sequence_numeric_values.append(20)
            
    return sequence_numeric_values

# Assigning the features and labels

X = []
y = sheets

lengths = []
for i in sequences:
    lengths.append(len(i))
max_len = max(lengths)
#counter = 0

for sequence in sequences:
    X.append(sequence_enhancer(sequence, max_len))
    #print(counter/8626*100, "% done")
    #counter= counter + 1

#print(len(X))


# Putting the features (positions) in the dataframe
X = pd.DataFrame(X)

#X = pd.concat([X, chains, length], ignore_index=True, axis=1)


X.head()

# logistic regression

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = 0.2)

clf = LogisticRegression(max_iter = 1000)     
clf.fit(X_train, y_train)
y_pred = clf.predict(X_test)

print(accuracy_score(y_pred, y_test))


# Checking the  logistic regression model using cross validation
cv_results = cross_validate(clf, X, y, cv=100)
print("Average accuracy is", np.average(cv_results["test_score"]))


######################################################################################################

# decision tree

treee = tree.DecisionTreeClassifier()
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2) 
treee = treee.fit(X_train, y_train)

y_pred = treee.predict(X_test)
#c_matrix = metrics.confusion_matrix(y_test, y_pred)
#print(c_matrix)
#print(" ")
print(accuracy_score(y_test, y_pred))  # accuracy for sheets

accuracy_list = []
for i in range(500):
    treee = tree.DecisionTreeClassifier()
    X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2) 
    treee = treee.fit(X_train, y_train)

    y_pred = treee.predict(X_test)
    accuracy_list.append(accuracy_score(y_test, y_pred))
    
sum(accuracy_list)/len(accuracy_list)

tree_results = cross_validate(treee, X, y, cv=50)
print(np.average(tree_results["test_score"]))

#########################################################################################################

# Try support vector machine

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = 0.3)
clf = svm.SVC(C = 3)
clf.fit(X_train, y_train)

y_pred = clf.predict(X_test)

print(accuracy_score(y_test, y_pred))
print(confusion_matrix(y_test, y_pred))

from sklearn.svm import SVC
svclassifier = SVC(kernel='linear')
svclassifier.fit(X_train, y_train)
y_pred = svclassifier.predict(X_test)
print(accuracy_score(y_test, y_pred))
