==> code/51 CSV convert sequences.ipynb <==
import pandas as pd

df = pd.read_csv("Data/many_proteins.csv")

df.head()

d3to1 = {'CYS': 'C', 'ASP': 'D', 'SER': 'S', 'GLN': 'Q', 'LYS': 'K',
 'ILE': 'I', 'PRO': 'P', 'THR': 'T', 'PHE': 'F', 'ASN': 'N', 
 'GLY': 'G', 'HIS': 'H', 'LEU': 'L', 'ARG': 'R', 'TRP': 'W', 
 'ALA': 'A', 'VAL':'V', 'GLU': 'E', 'TYR': 'Y', 'MET': 'M'}

sequences_list = []
translated_seq= ""
    
for sequence in df.sequences:
    sequence = sequence.replace("'", "")
    sequence = sequence.replace(" ", "")
    sequence = sequence.replace("[", "")
    sequence = sequence.replace("]", "")
    sequence = sequence.split(",")
    
    for j in sequence:
        translated_seq = translated_seq + d3to1[j]
        
    sequences_list.append(translated_seq)
    translated_seq= ""

df.drop("sequences", axis=1, inplace=True)
df.drop("original index", axis=1, inplace=True)
sequences = pd.DataFrame({'sequences': sequences_list})
df = pd.concat([df, sequences], axis=1)

df.to_csv("natural_seq.csv", index=False)

df.head()

# This code counts the incidence of each amino acid in each sequence and adds them to a list

df = pd.read_csv("natural_100_seq.csv")    

letters = ['C', 'D', 'S', 'Q', 'K', 'I', 'P', 'T', 'F', 'N', 'G', 'H', 'L', 'R', 'W', 'A', 'V', 'E', 'Y', 'M']
counts = {}

for i in letters:
    counts[i] = 0
    
C = []
D = []
S = []
Q = []
K = []
I = []
P = []
T = []
F = []
N = []
G = []
H = []
L = []
R = []
W = []
A = []
V = []
E = []
Y = []
M = []

#limit = 0
    
for sequence in df.sequences:
    #if limit == 50:
    #    break
    
    for letter in letters: # reset counts for each letter
        counts[letter] = 0

    #print("before", counts["C"])
        
    for acid in sequence:  # check for each acid in the sequence
        for letter in letters:
            if acid == letter:
                counts[letter] += 1

    #print("after", counts["C"])
    C.append(counts["C"])
    D.append(counts["D"])
    S.append(counts["S"])
    Q.append(counts["Q"])
    K.append(counts["K"])
    I.append(counts["I"])
    P.append(counts["P"])
    T.append(counts["T"])
    F.append(counts["F"])
    N.append(counts["N"])
    G.append(counts["G"])
    H.append(counts["H"])
    L.append(counts["L"])
    R.append(counts["R"])
    W.append(counts["W"])
    A.append(counts["A"])
    V.append(counts["V"])
    E.append(counts["E"])
    Y.append(counts["Y"])
    M.append(counts["M"])
    
    #limit += 1
    

df = pd.read_csv("natural_100_seq.csv")

zipped = list(zip(C, D, S, Q, K, I, P, T, F, N, G, H, L, R, W, A, V, E, Y, M))
counts = pd.DataFrame(zipped, columns=["C", "D", "S", "Q", "K", "I", "P", "T", "F", "N", "G", "H", "L", "R", "W", "A", "V", "E", "Y", "M"])

df = pd.concat([df, counts], axis=1)
df.head()


df.to_csv("natural_100_seq_feat.csv", index=False)

df = pd.read_csv("Data/natural_100_seq_feat.csv")
df.tail()
